# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BlihOps Talent & Employer Platform API
  description: |
    REST API for the BlihOps Talent & Employer Platform.
    
    This API provides endpoints for:
    - Talent profile management
    - Job posting and management
    - Talent-job matching
    - Administrative operations
    - File uploads (CVs)
    - Telegram webhook integration
    
    **Authentication**: All endpoints except `/auth/login` require JWT Bearer token authentication.
    
    **Base URL**: `https://api.blihops.com/api/v1`
    
    **Version**: 1.0.0
  version: 1.0.0
  contact:
    name: BlihOps API Support
    email: api-support@blihops.com
  license:
    name: Proprietary
    url: https://blihops.com/license
servers:
  - url: https://api.blihops.com/api/v1
    description: Production server
  - url: https://staging-api.blihops.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
paths:
...
  /talents:
    get:
      tags:
        - Talents
      summary: List talents
      description: Get paginated list of talents with filtering and sorting
      operationId: listTalents
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: Filter by approval status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: category
          in: query
          description: Filter by service category
          schema:
            $ref: '#/components/schemas/ServiceCategory'
        - name: skills
          in: query
          description: Filter by skills (comma-separated)
          schema:
            type: string
            example: TypeScript,NestJS,PostgreSQL
        - name: search
          in: query
          description: Search in name and skills
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, name, experienceLevel]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of talents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TalentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

    post:
      tags:
        - Talents
      summary: Create talent profile
      description: Create a new talent profile (typically called from Telegram bot)
      operationId: createTalent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTalentRequest'
      responses:
        '201':
          description: Talent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Talent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Talent with this Telegram ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /talents/{id}:
    get:
      tags:
        - Talents
      summary: Get talent by ID
      description: Get detailed talent profile
      operationId: getTalent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TalentIdParam'
      responses:
        '200':
          description: Talent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Talent'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Talents
      summary: Update talent profile
      description: Update talent profile (partial update)
      operationId: updateTalent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TalentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTalentRequest'
      responses:
        '200':
          description: Talent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Talent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Talents
      summary: Delete talent profile
      description: Soft delete talent profile
      operationId: deleteTalent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TalentIdParam'
      responses:
        '204':
          description: Talent deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /talents/{id}/approve:
    post:
      tags:
        - Talents
      summary: Approve talent profile
      description: Approve talent profile for public visibility
      operationId: approveTalent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TalentIdParam'
      responses:
        '200':
          description: Talent approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Talent'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Talent already approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /talents/{id}/reject:
    post:
      tags:
        - Talents
      summary: Reject talent profile
      description: Reject talent profile with optional reason
      operationId: rejectTalent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TalentIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Rejection reason
                  example: Profile incomplete
      responses:
        '200':
          description: Talent rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Talent'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    TelegramWebhookSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Secret token for Telegram webhook verification
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    TalentIdParam:
      name: id
      in: path
      required: true
      description: Talent UUID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Talent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        telegramId:
          type: integer
          description: Telegram user ID
          example: 123456789
        name:
          type: string
          example: John Doe
        serviceCategory:
          $ref: '#/components/schemas/ServiceCategory'
        roleSpecialization:
          type: string
          example: Senior Full-Stack Developer
        skills:
          type: array
          items:
            type: string
          example:
            - TypeScript
            - NestJS
            - PostgreSQL
            - React
        experienceLevel:
          $ref: '#/components/schemas/ExperienceLevel'
        availability:
          $ref: '#/components/schemas/Availability'
        engagementPreference:
          $ref: '#/components/schemas/EngagementType'
        cvFileId:
          type: string
          format: uuid
          nullable: true
          description: Reference to uploaded CV file
        status:
          $ref: '#/components/schemas/TalentStatus'
        consentForPublicVisibility:
          type: boolean
          default: true
        metadata:
          type: object
          description: Additional flexible data (JSONB)
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
          nullable: true
    CreateTalentRequest:
      type: object
      required:
        - telegramId
        - name
        - serviceCategory
        - skills
        - experienceLevel
        - availability
        - engagementPreference
      properties:
        telegramId:
          type: integer
          example: 123456789
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: John Doe
        serviceCategory:
          $ref: '#/components/schemas/ServiceCategory'
        roleSpecialization:
          type: string
          maxLength: 200
          example: Senior Full-Stack Developer
        skills:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 50
          example:
            - TypeScript
            - NestJS
            - PostgreSQL
        experienceLevel:
          $ref: '#/components/schemas/ExperienceLevel'
        availability:
          $ref: '#/components/schemas/Availability'
        engagementPreference:
          $ref: '#/components/schemas/EngagementType'
        cvFileId:
          type: string
          format: uuid
        consentForPublicVisibility:
          type: boolean
          default: true
        metadata:
          type: object
          additionalProperties: true
    UpdateTalentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        serviceCategory:
          $ref: '#/components/schemas/ServiceCategory'
        roleSpecialization:
          type: string
          maxLength: 200
        skills:
          type: array
          items:
            type: string
          maxItems: 50
        experienceLevel:
          $ref: '#/components/schemas/ExperienceLevel'
        availability:
          $ref: '#/components/schemas/Availability'
        engagementPreference:
          $ref: '#/components/schemas/EngagementType'
        cvFileId:
          type: string
          format: uuid
        consentForPublicVisibility:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
    TalentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Talent'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Jobs
...
    ServiceCategory:
      type: string
      enum:
        - ITO
        - AI_SOLUTIONS
        - AUTOMATION
        - DATA_ANALYTICS
      description: BlihOps service category
    ExperienceLevel:
      type: string
      enum:
        - JUNIOR
        - MID
        - SENIOR
        - LEAD
        - PRINCIPAL
      description: Experience level
    Availability:
      type: string
      enum:
        - IMMEDIATE
        - ONE_WEEK
        - TWO_WEEKS
        - ONE_MONTH
        - CUSTOM
      description: Availability status
    EngagementType:
      type: string
      enum:
        - FULL_TIME
        - PART_TIME
        - CONTRACT
        - FREELANCE
        - INTERNSHIP
      description: Engagement type
    TalentStatus:
      type: string
      enum:
        - pending
        - approved
        - rejected
      description: Talent profile approval status
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: Invalid input data
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /api/v1/talents
  responses:
    BadRequest:
      description: Bad request - validation error or invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: VALIDATION_ERROR
            message: Invalid input data
            details:
              field: email
              reason: Invalid email format
            timestamp: 2025-01-15T10:30:00Z
            path: /api/v1/talents
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHORIZED
            message: Invalid or expired token
            timestamp: 2025-01-15T10:30:00Z
            path: /api/v1/talents
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: Talent not found
            timestamp: 2025-01-15T10:30:00Z
            path: /api/v1/talents/123e4567-e89b-12d3-a456-426614174000
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per time window
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when rate limit resets
          example:
            error: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later.
            timestamp: 2025-01-15T10:30:00Z
            path: /api/v1/talents